+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## 1.0.3-SNAPSHOT
 
 * Strip nils from `:middleware`, fixes [#228](https://github.com/metosin/compojure-api/issues/228)
+* Lazily require `ring.swagger.validator` namespace in `compojure.api.swagger/validate` to allow compojure-api apps in [Google App Engine](https://cloud.google.com/appengine), Fixes [#227](https://github.com/metosin/compojure-api/issues/227). **NOTE** exluding `metosin/scjsv` will cause the validate to fail at runtime.
+
+```
+[metosin/compojure-api "1.0.3" :exclusions [[metosin/scjsv]]]
+```
 
 * updated dependencies: